JSCBC-850: Ensure all user-options are passed to bucket connections.
Motivation
----------
There was a bug in the original implementation of bucket-level connection
creation where not all the options that a user could specify were being
passed down and were instead only passed during cluster-level connection
creation.

Changes
-------
Merged the connection options creation between the two different kinds of
connections that we create to ensure all options were being appropriately
passed to both, and to prevent further divergence in the future.

Change-Id: I69da4b404083d0c698d4a49e933d8861480014fb